create template objects in wizard
