+++ src/utils/resources/template/utils/helpers.ts
@@ -4,12 +4,6 @@ import { V1Template } from '@kubevirt-ui/kubevirt-api/console';
 import VirtualMachineModel from '@kubevirt-ui/kubevirt-api/console/models/VirtualMachineModel';
 import { V1VirtualMachine } from '@kubevirt-ui/kubevirt-api/kubevirt';
 import { getAnnotation } from '@kubevirt-utils/resources/shared';
-import {
-  getGroupVersionKindForResource,
-  k8sCreate,
-  K8sModel,
-  K8sResourceCommon,
-} from '@openshift-console/dynamic-plugin-sdk';
 
 import { ANNOTATIONS } from './annotations';
 
@@ -39,27 +33,3 @@ export const replaceTemplateVM = (template: V1Template, vm: V1VirtualMachine) =>
     draftTemplate.objects.splice(vmIndex, 1, vm);
   });
 };
-
-export const createAllTemplateObjects = (
-  template: V1Template,
-  models: { [key: string]: K8sModel },
-  namespace?: string,
-): Promise<K8sResourceCommon[]> =>
-  Promise.all(
-    template.objects.map((object) => {
-      const { group, version, kind } = getGroupVersionKindForResource(object);
-
-      const ref = [group || 'core', version, kind].join('~');
-
-      const model = models[ref] || models[object.kind];
-
-      return k8sCreate<K8sResourceCommon>({
-        model: model,
-        data: produce(object as K8sResourceCommon, (draftObject) => {
-          if (!draftObject.metadata.namespace && namespace) {
-            draftObject.metadata.namespace = namespace;
-          }
-        }),
-      });
-    }),
-  );
